+++ dflib-jdbc/src/test/java/com/nhl/dflib/jdbc/connector/JdbcConnectorBuilder_IT.java
@@ -1,34 +1,39 @@
 package com.nhl.dflib.jdbc.connector;
 
+import com.nhl.dflib.DataFrame;
+import com.nhl.dflib.jdbc.Jdbc;
+import com.nhl.dflib.jdbc.connector.loader.ColumnBuilderFactory;
 import com.nhl.dflib.jdbc.unit.BaseDbTest;
+import com.nhl.dflib.unit.DataFrameAsserts;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 public class JdbcConnectorBuilder_IT extends BaseDbTest {
 
     @Test
-    public void testBuild() throws SQLException {
+    @DisplayName("Connector smoke test")
+    public void testSmoke() throws SQLException {
 
-        JdbcConnector connector = adapter.createConnector();
+        JdbcConnector connector = Jdbc.connector().dataSource(adapter.getDb().getDataSource()).build();
+        String table = "JdbcConnectorBuilder_IT_smoke_test";
 
-        String sql = adapter.toNativeSql("create table \"JdbcConnectorBuilder_IT\" (\"id\" int)");
-
+        String sql = adapter.toNativeSql("create table \"%s\" (\"id\" int)", table);
         try (Connection c1 = connector.getConnection()) {
             try (PreparedStatement st = c1.prepareStatement(sql)) {
-               st.executeUpdate();
+                st.executeUpdate();
             }
 
             c1.commit();
         }
 
-        sql = adapter.toNativeSql("insert into \"JdbcConnectorBuilder_IT\" values (356)");
-
+        sql = adapter.toNativeSql("insert into \"%s\" values (356)", table);
         try (Connection c2 = connector.getConnection()) {
             try (PreparedStatement st = c2.prepareStatement(sql)) {
                 st.executeUpdate();
@@ -37,8 +42,7 @@ public class JdbcConnectorBuilder_IT extends BaseDbTest {
             c2.commit();
         }
 
-        sql = adapter.toNativeSql("select * from \"JdbcConnectorBuilder_IT\"");
-
+        sql = adapter.toNativeSql("select * from \"%s\"", table);
         try (Connection c3 = connector.getConnection()) {
             try (PreparedStatement st = c3.prepareStatement(sql)) {
                 try (ResultSet rs = st.executeQuery()) {
@@ -51,4 +55,51 @@ public class JdbcConnectorBuilder_IT extends BaseDbTest {
             c3.commit();
         }
     }
+
+    @Test
+    @DisplayName("Custom ColumnBuilderFactories")
+    public void testAddColumnBuilderFactory() throws SQLException {
+
+        String table = "JdbcConnectorBuilder_IT_AddColumnBuilderFactory";
+        String sql = adapter.toNativeSql("create table \"%s\" (\"id\" bigint, \"d\" date, \"t\" time, \"ts\" timestamp)", table);
+
+        try (Connection c1 = adapter.createConnector().getConnection()) {
+            try (PreparedStatement st = c1.prepareStatement(sql)) {
+                st.executeUpdate();
+            }
+
+            c1.commit();
+        }
+
+        LocalDate d = LocalDate.of(1977, 02, 05);
+        LocalTime t = LocalTime.of(5, 6, 8);
+        LocalDateTime ts = LocalDateTime.of(2019, 02, 03, 1, 2, 5);
+
+        adapter.getTable(table).insertColumns("id", "d", "t", "ts").values(1L, d, t, ts).exec();
+
+        DataFrame customTypes = Jdbc.connector()
+                .dataSource(adapter.getDb().getDataSource())
+                .addColumnBuilderFactory(Types.DATE, ColumnBuilderFactory::objectAccum)
+                .addColumnBuilderFactory(Types.TIME, ColumnBuilderFactory::objectAccum)
+                .addColumnBuilderFactory(Types.TIMESTAMP, ColumnBuilderFactory::objectAccum)
+                .build()
+                .tableLoader(table)
+                .load();
+
+        new DataFrameAsserts(customTypes, "id", "d", "t", "ts")
+                .expectHeight(1)
+                .expectRow(0, 1L, Date.valueOf(d), Time.valueOf(t), Timestamp.valueOf(ts));
+
+        DataFrame standardTypes = Jdbc.connector()
+                .dataSource(adapter.getDb().getDataSource())
+                .build()
+                .tableLoader(table)
+                .load();
+
+        new DataFrameAsserts(standardTypes, "id", "d", "t", "ts")
+                .expectHeight(1)
+                .expectRow(0, 1L, d, t, ts);
+    }
+
+
 }