+++ dflib-jdbc/src/test/java/com/nhl/dflib/jdbc/unit/dbadapter/GenericTestAdapter.java
@@ -16,7 +16,7 @@ public class GenericTestAdapter implements TestDbAdapter {
     }
 
     @Override
-    public String toNativeSql(String command) {
-        return command;
+    public String toNativeSql(String derbySql, String... params) {
+        return String.format(derbySql, params);
     }
 }