+++ dflib-jdbc/src/test/java/com/nhl/dflib/jdbc/unit/dbadapter/TestDbAdapter.java
@@ -15,7 +15,7 @@ public interface TestDbAdapter {
 
     DbTester getDb();
 
-    String toNativeSql(String derbySql);
+    String toNativeSql(String derbySql, String... params);
 
     default JdbcConnector createConnector() {
         return Jdbc.connector(getDb().getDataSource());